# api/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/wg/peers": {
            "get": {
                "description": "Admin: list all peers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wg"
                ],
                "summary": "List WireGuard peers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by user id",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by device name (contains)",
                        "name": "device_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by client ip (exact)",
                        "name": "client_ip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (active/revoked)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Peers listed successfully",
                        "schema": {
                            "$ref": "#/definitions/v1.WGPeerListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Admin: create peer (auto generate keys/ip, write files, apply)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wg"
                ],
                "summary": "Create WireGuard peer",
                "parameters": [
                    {
                        "description": "Create peer payload",
                        "name": "peer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreateWGPeerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Peer created successfully",
                        "schema": {
                            "$ref": "#/definitions/v1.WGPeerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/wg/peers/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wg"
                ],
                "summary": "Get WireGuard peer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Peer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Peer retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/v1.WGPeerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wg"
                ],
                "summary": "Update WireGuard peer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Peer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update peer payload",
                        "name": "peer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.UpdateWGPeerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Peer updated successfully",
                        "schema": {
                            "$ref": "#/definitions/v1.WGPeerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wg"
                ],
                "summary": "Delete (revoke) WireGuard peer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Peer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Peer revoked successfully",
                        "schema": {
                            "$ref": "#/definitions/core.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "core.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code defines the business error code.",
                    "type": "integer"
                },
                "details": {
                    "description": "Details contains detailed validation error information.\nThis field is only present when validation errors occur.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "description": "Message contains the detail of this message.\nThis message is suitable to be exposed to external",
                    "type": "string"
                },
                "reference": {
                    "description": "Reference returns the reference document which maybe useful to solve this error.",
                    "type": "string"
                }
            }
        },
        "core.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code defines the business success code.",
                    "type": "integer"
                },
                "message": {
                    "description": "Message contains the success message.",
                    "type": "string"
                }
            }
        },
...
        "v1.CreateWGPeerRequest": {
            "type": "object",
            "required": [
                "device_name",
                "username"
            ],
            "properties": {
                "allowed_ips": {
                    "description": "AllowedIPs optionally overrides server-side AllowedIPs for this peer.",
                    "type": "string",
                    "maxLength": 512
                },
                "device_name": {
                    "description": "DeviceName is a human friendly device name.",
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1
                },
                "persistent_keepalive": {
                    "description": "PersistentKeepalive is optional keepalive in seconds (0 means unset).",
                    "type": "integer",
                    "maximum": 3600,
                    "minimum": 0
                },
                "username": {
                    "description": "Username is the owner username.",
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 3
                }
            }
        },
...
        "v1.UpdateWGPeerRequest": {
            "type": "object",
            "properties": {
                "allowed_ips": {
                    "type": "string",
                    "maxLength": 512
                },
                "persistent_keepalive": {
                    "type": "integer",
                    "maximum": 3600,
                    "minimum": 0
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "revoked"
                    ]
                }
            }
        },
...
        "v1.WGPeerListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.WGPeerResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "v1.WGPeerResponse": {
            "type": "object",
            "properties": {
                "allowed_ips": {
                    "type": "string"
                },
                "client_ip": {
                    "type": "string"
                },
                "client_public_key": {
                    "type": "string"
                },
                "device_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "persistent_keepalive": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}